+++ WeBASE-Data-Collect/src/main/java/com/webank/webase/data/collect/base/enums/GasRecordType.java
@@ -11,19 +11,33 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.webank.webase.data.collect.config.entity;
 
-import java.time.LocalDateTime;
-import lombok.Data;
+package com.webank.webase.data.collect.base.enums;
 
-@Data
-public class TbConfigVersion {
-    private Integer id;
-    private String chainInfoVersion;
-    private String frontInfoVersion;
-    private String groupInfoVersion;
-    private String userInfoVersion;
-    private String contractInfoVersion;
-    private LocalDateTime createTime;
-    private LocalDateTime modifyTime;
+/**
+ * Enumeration of gas record type.
+ */
+public enum GasRecordType {
+    consume(0), charge(1), deduct(2);
+
+    private int type;
+
+    private GasRecordType(int type) {
+        this.type = type;
+    }
+
+    public int getType() {
+        return this.type;
+    }
+
+    public static boolean isInclude(int type) {
+        boolean include = false;
+        for (GasRecordType e : GasRecordType.values()) {
+            if (e.getType() == type) {
+                include = true;
+                break;
+            }
+        }
+        return include;
+    }
 }