+++ WeBASE-Data-Collect/src/main/java/com/webank/webase/data/collect/base/properties/ConstantProperties.java
@@ -29,7 +29,13 @@ public class ConstantProperties {
     public static final String CONTRACT_NAME_ZERO = "0x00000000";
     public static final String ADDRESS_DEPLOY = "0x0000000000000000000000000000000000000000";
     
+    public static final String TB_GAS = "tb_gas";
+    public static final String PREFIX_PARTITION = "p_";
     public static final String ORDER_BY_ID_DESC = "id DESC";
+    
+    // gas parser
+    public static final String GAS_CHARGE = "charge(address,uint256)";
+    public static final String GAS_DEDUCT = "deduct(address,uint256)";
 
     // group info
     private String groupInvalidGrayscaleValue = "1M"; // y:year, M:month, d:day of month, h:hour,
@@ -38,7 +44,6 @@ public class ConstantProperties {
     // data pull
     private long startBlockNumber = 0;
     private int crawlBatchUnit = 100;
-    private Long pullBlockSleepTime1 = 20L; // 20 mills
 
     // front http request
     private Integer httpTimeOut = 5000;
@@ -47,11 +52,10 @@ public class ConstantProperties {
 
     // data handle
     private boolean ifPullData = true; // default true
+    private boolean ifSaveBlockAndTrans = true; // default true
+    private boolean ifSaveGas = true; // default true
+    private boolean ifEsEnable = false; // default false
     private boolean multiLiving = false; // default false
     private Long dataParserTaskFixedDelay = 60000L; // default 1min
-    
-    // config server
-    private boolean syncConfig = false; // default false
-    private String configServerIpPort = "127.0.0.1:8001";
 
 }