+++ WeBASE-Data-Collect/src/main/java/com/webank/webase/data/collect/user/UserService.java
@@ -15,7 +15,6 @@ package com.webank.webase.data.collect.user;
 
 import com.webank.webase.data.collect.base.code.ConstantCode;
 import com.webank.webase.data.collect.base.exception.BaseException;
-import com.webank.webase.data.collect.config.entity.RspUserInfo;
 import com.webank.webase.data.collect.group.GroupService;
 import com.webank.webase.data.collect.parser.ParserService;
 import com.webank.webase.data.collect.user.entity.TbUser;
@@ -28,7 +27,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.CollectionUtils;
 
 /**
  * services for user data.
@@ -44,17 +42,6 @@ public class UserService {
     private ParserService parserService;
     @Autowired
     private GroupService groupService;
-    
-    /**
-     * add user list
-     */
-    public void addUserList(List<RspUserInfo> userList) {
-        if (CollectionUtils.isEmpty(userList)) {
-            log.info("contractList is empty.");
-            return;
-        }
-        userMapper.addList(userList);
-    }
 
     /**
      * add user info.