+++ WeBASE-Data-Fetcher/Changelog.md
@@ -1,3 +1,24 @@
+### v1.1.0 (2021-01-19)
+
+​	WeBASE-Data-Fetcher（微众区块链中间件平台-数据监管服务数据读取子系统）。
+
+**Add**
+
+- 增加Elasticsearch开关配置ifEsEnable，为false时不支持数据检索
+
+**Update**
+
+- jar包更新
+
+**兼容性**
+
+- 支持FISCO-BCOS 2.3.0或以上版本
+- WeBASE-Front 1.3.0或以上版本
+
+详细了解,请阅读[**技术文档**](https://webasedoc.readthedocs.io/zh_CN/latest/)
+
+
+
 ### v1.0.0 (2020-06-20)
 
 ​	WeBASE-Data-Fetcher（微众区块链中间件平台-数据监管服务数据读取子系统）。